add actiontext to exercices and change some strings types for text.

+++ db/migrate/20191224144237_create_languages.rb
@@ -2,7 +2,7 @@ class CreateLanguages < ActiveRecord::Migration[6.0]
   def change
     create_table :languages do |t|
       t.string :name
-      t.numeric :cod
+      t.integer :cod
 
       t.timestamps
     end
